# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oso Frontino API",
        "description": "API for iPhone store - Products, Baskets and Invoices. Use POST /auth/login or POST /auth/register to get a Bearer token. Include it in the Authorization header for create, update and delete operations.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API (relative - same origin)"
        },
        {
            "url": "http://127.0.0.1:8002/api",
            "description": "Local - 127.0.0.1"
        },
        {
            "url": "http://localhost:8002/api",
            "description": "Local - localhost"
        }
    ],
    "paths": {
...
        "/androids/{id}": {
            "get": {
                "tags": [
                    "Androids"
                ],
                "summary": "Get android by ID",
                "operationId": "androids.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Androids"
                ],
                "summary": "Update android",
                "operationId": "androids.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Androids"
                ],
                "summary": "Delete android",
                "operationId": "androids.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/audifonos/{id}": {
            "get": {
                "tags": [
                    "Audifonos"
                ],
                "summary": "Get audifono by ID",
                "operationId": "audifonos.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Audifonos"
                ],
                "summary": "Update audifono",
                "operationId": "audifonos.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Audifonos"
                ],
                "summary": "Delete audifono",
                "operationId": "audifonos.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/baskets/{id}": {
            "get": {
                "tags": [
                    "Baskets"
                ],
                "summary": "Get basket by ID",
                "operationId": "baskets.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Baskets"
                ],
                "summary": "Update basket",
                "operationId": "baskets.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Baskets"
                ],
                "summary": "Delete basket",
                "operationId": "baskets.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/iphones/{id}": {
            "get": {
                "tags": [
                    "iPhones"
                ],
                "summary": "Get iPhone by ID",
                "operationId": "iphones.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "iPhone details",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "name": "iPhone 15 Pro",
                                        "description": "Latest iPhone model",
                                        "price": 999.99,
                                        "discount": 10,
                                        "quantityStock": 5,
                                        "amount": 899.99,
                                        "imagePath": "storage/iphones/iphone15.jpg",
                                        "createdAt": "2025-02-23T10:00:00+00:00",
                                        "updatedAt": "2025-02-23T10:00:00+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "iPhones"
                ],
                "summary": "Update iPhone",
                "operationId": "iphones.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "iPhone 15 Pro Max",
                                "description": "Updated description",
                                "price": 1099.99,
                                "discount": 15,
                                "quantity_stock": 10,
                                "amount": 934.99,
                                "image_path": "storage/iphones/iphone15pro.jpg"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "iPhone updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "iPhones"
                ],
                "summary": "Delete iPhone",
                "operationId": "iphones.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/smart-watches": {
            "get": {
                "tags": [
                    "Smart Watches"
                ],
                "summary": "List all smart watches",
                "operationId": "smartWatches.index",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Smart Watches"
                ],
                "summary": "Create smart watch",
                "operationId": "smartWatches.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ]
                            },
                            "example": {
                                "name": "Apple Watch",
                                "price": 399.99
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/smart-watches/{id}": {
            "get": {
                "tags": [
                    "Smart Watches"
                ],
                "summary": "Get smart watch by ID",
                "operationId": "smartWatches.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Smart Watches"
                ],
                "summary": "Update smart watch",
                "operationId": "smartWatches.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Smart Watches"
                ],
                "summary": "Delete smart watch",
                "operationId": "smartWatches.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Enter your Bearer token from login/register"
            }
        }
    },
...
}